Project Overview :
Problem : The manual system is suffering from a series of drawbaks. Since whole of the bills is to be maintained with hands the process of keeping and maintaining the information is very tedious and lengthy to customer. It is very time consuming and laborious process because, staff need to be visted the customers place ever month togive the bills and to receive the payments. For thes reason, we have provided features present system is partially automated, existing system is quite laborious as one must enter same information at different places.

Solution :  This project system excludes the need of maintaining paper electricity bill as all the electricity bill records are managed electronically. Administrator diesnt have to visit to the office for bill payment. There is no need of delivery boy for dekuvert bills ti users place. Thus, it saves human efforts and resources.